+++ src/sonar.rs
@@ -15,10 +15,57 @@
 //! Finding C declarations.
 
 use std::collections::{HashSet};
+use std::str::{FromStr};
 
 use super::{Entity, EntityKind, Type, TypeKind};
 
 type Seen = HashSet<String>;
+
+//================================================
+// Enums
+//================================================
+
+// DefinitionValue _______________________________
+
+/// The value of a C preprocessor definition.
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum DefinitionValue {
+    /// An integer.
+    Integer(bool, u64),
+    /// A floating point number.
+    Real(f64),
+}
+
+impl DefinitionValue {
+    //- Constructors -----------------------------
+
+    fn from_entity(entity: Entity) -> Option<DefinitionValue> {
+        let mut tokens = entity.get_range().unwrap().tokenize();
+        if tokens.last().map_or(false, |t| t.get_spelling() == "#") {
+            tokens.pop();
+        }
+
+        let (negated, number) = if tokens.len() == 2 {
+            (false, tokens[1].get_spelling())
+        } else if tokens.len() == 3 && tokens[1].get_spelling() == "-" {
+            (true, tokens[2].get_spelling())
+        } else {
+            return None;
+        };
+
+        if let Ok(integer) = u64::from_str(&number) {
+            Some(DefinitionValue::Integer(negated, integer))
+        } else if let Ok(real) = f64::from_str(&number) {
+            if negated {
+                Some(DefinitionValue::Real(-real))
+            } else {
+                Some(DefinitionValue::Real(real))
+            }
+        } else {
+            None
+        }
+    }
+}
 
 //================================================
 // Structs
@@ -42,6 +89,27 @@ impl<'tu> Declaration<'tu> {
 
     fn new(name: String, entity: Entity<'tu>, source: Option<Entity<'tu>>) -> Declaration<'tu> {
         Declaration { name: name, entity: entity, source: source }
+    }
+}
+
+// Definition ____________________________________
+
+/// A C preprocessor definition.
+#[derive(Clone, Debug, PartialEq)]
+pub struct Definition<'tu> {
+    /// The name of the definition.
+    pub name: String,
+    /// The value of the definition.
+    pub value: DefinitionValue,
+    /// The entity that describes the definition.
+    pub entity: Entity<'tu>,
+}
+
+impl<'tu> Definition<'tu> {
+    //- Constructors -----------------------------
+
+    fn new(name: String, value: DefinitionValue, entity: Entity<'tu>) -> Definition<'tu> {
+        Definition { name: name, value: value, entity: entity }
     }
 }
 
@@ -100,7 +168,39 @@ fn visit<'tu, F: FnMut(Declaration<'tu>)>(
     }
 }
 
-/// Finds the enums in the supplied entities.
+/// Visits the simple preprocessor definitions in the supplied entities.
+///
+/// Simple preprocessor definitions are those that consist only of a single integer or floating
+/// point literal, optionally negated.
+///
+/// If a preprocessor definition is encountered multiple times, only the first instance is visited.
+#[cfg_attr(feature="clippy", allow(needless_lifetimes))]
+pub fn visit_definitions<'tu, F: FnMut(Definition<'tu>)>(entities: &[Entity<'tu>], mut f: F) {
+    let mut seen = Seen::new();
+
+    for entity in entities.iter().filter(|e| e.get_kind() == EntityKind::MacroDefinition) {
+        let name = entity.get_name().unwrap();
+        let range = entity.get_range().unwrap();
+
+        if !seen.contains(&name) && !range.get_start().is_in_system_header() {
+            if let Some(value) = DefinitionValue::from_entity(*entity) {
+                f(Definition::new(name.clone(), value, *entity));
+                seen.insert(name);
+            }
+        }
+    }
+}
+
+/// Finds the simple preprocessor definitions in the supplied entities.
+///
+/// See `visit_definitions` for more information.
+pub fn find_definitions<'tu>(entities: &[Entity<'tu>]) -> Vec<Definition<'tu>> {
+    let mut definitions = vec![];
+    visit_definitions(entities, |d| definitions.push(d));
+    definitions
+}
+
+/// Returns the enums in the supplied entities.
 ///
 /// If an enum is encountered multiple times, only the first instance is collected.
 pub fn find_enums<'tu>(entities: &[Entity<'tu>]) -> Vec<Declaration<'tu>> {
@@ -117,7 +217,7 @@ pub fn visit_enums<'tu, F: FnMut(Declaration<'tu>)>(entities: &[Entity<'tu>], f:
     visit(entities, f, EntityKind::EnumDecl, "enum ");
 }
 
-/// Finds the functions in the supplied entities.
+/// Returns the functions in the supplied entities.
 ///
 /// If a function is encountered multiple times, only the first instance is collected.
 pub fn find_functions<'tu>(entities: &[Entity<'tu>]) -> Vec<Declaration<'tu>> {
@@ -143,7 +243,7 @@ pub fn visit_functions<'tu, F: FnMut(Declaration<'tu>)>(entities: &[Entity<'tu>]
     }
 }
 
-/// Finds the structs in the supplied entities.
+/// Returns the structs in the supplied entities.
 ///
 /// If a struct is encountered multiple times, only the first instance is collected.
 pub fn find_structs<'tu>(entities: &[Entity<'tu>]) -> Vec<Declaration<'tu>> {
@@ -160,7 +260,7 @@ pub fn visit_structs<'tu, F: FnMut(Declaration<'tu>)>(entities: &[Entity<'tu>], 
     visit(entities, f, EntityKind::StructDecl, "struct ");
 }
 
-/// Finds the typedefs in the supplied entities.
+/// Returns the typedefs in the supplied entities.
 ///
 /// If a typedef is encountered multiple times, only the first instance is collected.
 pub fn find_typedefs<'tu>(entities: &[Entity<'tu>]) -> Vec<Declaration<'tu>> {
@@ -195,7 +295,7 @@ pub fn visit_typedefs<'tu, F: FnMut(Declaration<'tu>)>(entities: &[Entity<'tu>],
     }
 }
 
-/// Finds the unions in the supplied entities.
+/// Returns the unions in the supplied entities.
 ///
 /// If a union is encountered multiple times, only the first instance is collected.
 pub fn find_unions<'tu>(entities: &[Entity<'tu>]) -> Vec<Declaration<'tu>> {